authorization check when editing and deleting

+++ controllers/feed.js
@@ -144,6 +144,12 @@ exports.updatePost = async (req, res, next) => {
             throw error;
         }
 
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('not authorized');
+            error.statusCode = 403;
+            throw error;
+        }
+
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -179,6 +185,13 @@ exports.deletePost = async (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('not authorized');
+            error.statusCode = 403;
+            throw error;
+        }
+
         clearImage(post.imageUrl);
 
         const removedPost = await Post.findByIdAndDelete(postId);
